feat(sdk): updates all sdks with latest versions
