+++ cli/src/command/init/template/go/Vorpal.lock
@@ -1,0 +1,16 @@
+{
+    "alias": {
+        "AARCH64_DARWIN": {
+            "go:1.24.2": "94441629fea5e77152d14134329abe04000f0153b6bbdd43d50ae01082044d97"
+        },
+        "AARCH64_LINUX": {
+            "go:1.24.2": "fa75f3e77a88a633c74d457e17f6fcd2c9ed80f0655d29d5fcb2a39308cc540b"
+        },
+        "X8664_DARWIN": {
+            "go:1.24.2": "d05220d6236a082594f8122154b656c29b7791f183d467732c21eecb28feb592"
+        },
+        "X8664_LINUX": {
+            "go:1.24.2": "7e76e64ed9e1b1187bb83e0ce700e38a9fd7917f65b8b3523506b659ce1849c9"
+        }
+    }
+}
