+++ cli/src/command/init/template/go/main.go
