+++ cli/src/command/init/template/rust/vorpal/src/main.rs
@@ -1,12 +1,22 @@
 use anyhow::Result;
-use vorpal_sdk::{artifact::language::rust::RustBuilder, context::get_context};
+use vorpal_sdk::{
+    api::artifact::{
+        ArtifactSystem,
+        ArtifactSystem::{Aarch64Darwin, Aarch64Linux, X8664Darwin, X8664Linux},
+    },
+    artifact::language::rust::RustBuilder,
+    context::get_context,
+};
+
+const SYSTEMS: [ArtifactSystem; 4] = [Aarch64Darwin, Aarch64Linux, X8664Darwin, X8664Linux];
 
 #[tokio::main]
 async fn main() -> Result<()> {
     let context = &mut get_context().await?;
 
-    RustBuilder::new("example")
+    RustBuilder::new("example", SYSTEMS.to_vec())
         .with_bins(vec!["example"])
+        .with_includes(vec!["example", "Cargo.lock", "Cargo.toml"])
         .with_packages(vec!["example"])
         .build(context)
         .await?;
